chinese_chr_app/chinese_chr_app/backend/requirements.txt
Flask==3.0.0
psycopg[binary]>=3.1
flask-cors==4.0.0
gunicorn==21.2.0
google-cloud-storage==2.10.0
python-dotenv==1.0.0
certifi==2026.1.4
PyJWT[crypto]==2.10.1